www/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
mysql
mysql-connector-python
mysqlclient
PyMySQL
python-dotenv
Scrapy
SQLAlchemy